requirements.txt
streamlit
crewai
python-dotenv
onnxruntime
langchain-community